+++ src/renderer/services/docker/stats.ts
